feat: entity-place; 장소조회 add pinCnt

+++ src/main/java/com/pintogether/backend/dto/PlaceResponseDTO.java
@@ -13,7 +13,7 @@ public class PlaceResponseDTO {
 
     private String roadNameAddress;
 
-    private int pinCnt;
+    private Long pinCnt;
 
     private double latitude;
